imagedownloader/requirements.txt
Beaker==1.5.4
Django==1.5
IPy==0.75
PyVirtualDisplay==0.1.1
PyYAML==3.10
South==0.7.6
decorator==3.4.0
#distribute==0.6.27
django-model-utils==1.2.0
docutils==0.8.1
email==4.0.2
numpy==1.8.0
h5py==2.1.2
ipython==0.12
mglob==0.4
netCDF4==1.0.2
python-dateutil==1.5
pytz==2012j
#scipy==0.10.1
splinter==0.4.9
pyttsx==1.1
django-polymorphic==0.5.3
pyquery==1.2.6
xlrd==0.9.2
coveralls==0.3
django-sslify==0.2.2
defusedxml==0.4.1
lxml==2.3.6
django-tastypie==0.11.0